+++ tests/testthat/test_exp.R
@@ -1,6 +1,8 @@
 context("Exposure")
 
 # Set params
+data <- sales
+data$region <- as.factor(data$region)
 Ptime <- "sales_month"
 Pdevice_hierarchy <- "device_name"
 Pevent_hierarchy <- NULL
@@ -10,7 +12,7 @@ Pcount <- "sales_volume"
 
 # Reference example
 a1 <- exposure(
-  sales,
+  data,
   time=Ptime,
   device_hierarchy=Pdevice_hierarchy,
   match_levels=Pmatch_levels,
@@ -28,13 +30,13 @@ test_that("function returns the correct class", {
 })
 test_that("parameter requirements as expected", {
   expect_error(exposure())
-  expect_error(exposure(sales))
-  expect_error(exposure(sales, Ptime))
-  expect_is(exposure(sales, Ptime, Pdevice_hierarchy), "mds_e")
-  expect_error(exposure(sales, Ptime, Pdevice_hierarchy, event_hierarchy="foo"))
-  expect_error(exposure(sales, Ptime, Pdevice_hierarchy, key="foo"))
-  expect_error(exposure(sales, Ptime, Pdevice_hierarchy, match_levels="foo"))
-  expect_error(exposure(sales, Ptime, Pdevice_hierarchy, count="foo"))
+  expect_error(exposure(data))
+  expect_error(exposure(data, Ptime))
+  expect_is(exposure(data, Ptime, Pdevice_hierarchy), "mds_e")
+  expect_error(exposure(data, Ptime, Pdevice_hierarchy, event_hierarchy="foo"))
+  expect_error(exposure(data, Ptime, Pdevice_hierarchy, key="foo"))
+  expect_error(exposure(data, Ptime, Pdevice_hierarchy, match_levels="foo"))
+  expect_error(exposure(data, Ptime, Pdevice_hierarchy, count="foo"))
 })
 # Attribute check
 test_that("attributes are fully described", {
@@ -53,18 +55,18 @@ test_that("attributes are fully described", {
 # ------------------------
 
 test_that("output shape is as expected", {
-  expect_equal(nrow(a1), nrow(sales))
-  expect_equal(ncol(a1), ncol(sales) + 1)
-  expect_equal(sum(is.na(a1[[Pmatch_levels]])), sum(is.na(sales[[Pmatch_levels]])))
-  expect_equal(sum(is.na(a1$device_1)), sum(is.na(sales[[Pdevice_hierarchy]])))
+  expect_equal(nrow(a1), nrow(data))
+  expect_equal(ncol(a1), ncol(data) + 1)
+  expect_equal(sum(is.na(a1[[Pmatch_levels]])), sum(is.na(data[[Pmatch_levels]])))
+  expect_equal(sum(is.na(a1$device_1)), sum(is.na(data[[Pdevice_hierarchy]])))
 })
 
 test_that("input variable matches mapped output variable", {
-  expect_equal(as.character(a1$key), as.character(c(1:nrow(sales))))
-  expect_equal(a1$time, sales[[Ptime]])
-  expect_equal(a1$count, sales[[Pcount]])
-  expect_equal(as.character(a1[[Pmatch_levels]]), sales[[Pmatch_levels]])
-  expect_equal(as.character(a1$device_1), sales[[Pdevice_hierarchy]])
+  expect_equal(as.character(a1$key), as.character(c(1:nrow(data))))
+  expect_equal(a1$time, data[[Ptime]])
+  expect_equal(a1$count, data[[Pcount]])
+  expect_equal(a1[[Pmatch_levels]], data[[Pmatch_levels]])
+  expect_equal(as.character(a1$device_1), data[[Pdevice_hierarchy]])
 })
 
 test_that("output variable class converted correctly", {
@@ -83,7 +85,7 @@ test_that("no extra variables were created", {
 
 test_that("descriptors were kept in source format", {
   expect_is(class(a1$key), "character")
-  expect_equal(class(a1$count), class(sales[[Pcount]]))
+  expect_equal(class(a1$count), class(data[[Pcount]]))
 })
 
 # Barebones behavior
@@ -95,15 +97,15 @@ Pdevice_hierarchy <- "device_name"
 
 # Reference example
 a1 <- exposure(
-  sales,
+  data,
   time=Ptime,
   device_hierarchy=Pdevice_hierarchy
 )
 
 test_that("minimal parameters output shape is as expected", {
-  expect_equal(nrow(a1), nrow(sales))
-  expect_equal(ncol(a1), ncol(sales))
-  expect_equal(a1$key, as.character(c(1:nrow(sales))))
+  expect_equal(nrow(a1), nrow(data))
+  expect_equal(ncol(a1), ncol(data))
+  expect_equal(a1$key, as.character(c(1:nrow(data))))
   expect_null(a1$region)
 })